Add Category and Sourcing admin pages and remove user list page test
The user list page is now restricted to the new "framework manager" role
which we don't have a user account set up for yet.  And the page will be
going away altogether in the next week or so, so not worth patching up
that test right now.

Category and Sourcing accounts added, with one page tested for each new
type of user.