Re-order test scenarios to provoke new login
**THIS IS A TERRIBLE HACK THAT MUST BE FIXED PROPERLY SOON**

The tests are deciding whether to re-log in based on if the bot is redirected to the login page when it tries to access the URL.  This meant that for different admin user roles the new role was not being logged in, and all admin pages screenshotted as admin (support) role.

By rearranging so that admin roles are interleaved with other roles we should be able to provoke a new login with the correct user for this set of tests.

This is by no means a permanent solution, but a quick hack for Friday to unblock our release pipeline.  Someone *will* make a better fix soon.