Update jenkins UID and bump verison to 2.0.0
The visual regression tests Dockerfile creates a user called 'jenkins' and specifies it's UID. This UID has to match with the owner of the directory that gets mounted as a volume into the container. Currently the UID in the Dockerfile is hardcoded to match the UID of the jenkins user on the Jenkins box. This is bad. If the jenkins user changes UID (maybe the box gets rebuilt again) the job breaks.