Use Google Chrome and Puppeteer for visual regression testing
PhantomJS has been abandoned and BackstopJS is dropping support for it,
so we should use the new default engine, puppeteer with chrome.

This commit adds engine_scripts for puppeteer to give the same
functionality as we used with PhantomJS. The starting point for these
scripts were the ones shipped with backstopjs in the
`node_modules/backstopjs/capture/engine_scripts/puppet` folder, and then
I ported the features from our own `backstop_data/engine_scripts/casper`
scripts.