Fix login script
As part of migrating our frontends to use GOV.UK Design System
components, we replaced the buttons in our user frontend (see
alphagov/digitalmarketplace-user-frontend#167).

This also had the effect of breaking our visual regression tests,
because our script to login to the frontend was looking for the old
button class.

This commit changes the script to work with both old and new style
buttons.