Add more admin user roles
We want to test pages when logged in as admin managers, framework
manager admins, and CCS data controller admins.

This commit adds extra tests to the `user.js` engine script to allow
logging in as these users.

The Jenkins job for the visual regression tests already includes the
required environment variables[1], so this should work on Jenkins
without any other PRs required.

[1]: https://github.com/alphagov/digitalmarketplace-jenkins/blob/master/job_definitions/visual_regression.yml#L69-L93