Increase specificity of login button selector
We have had an issue where our visual regression tests have been
clicking on the coookie banner accept cookies button instead of the
login button, as both buttons use the same class.

In the long term we should consider adding an id to the login button to
make it easier to select (or an id to the cookie banner buttons to make
it easier to NOT select them). For now ensuring that the button is
inside the main wrapper is sufficient to make the tests do what we want.