Wait for fonts to load and scroll to the top
On backstopjs 5.1 and above we were seeing test failures with very
minor rendering differences to the reference images, and also fonts
not always loading.

We are not the only people seeing these problems - see [1] and [2].
In those issues, the author of backstopjs suggests that this is
actually an issue with puppeteer, and another user suggests adding
these lines to onReady.js to force the engine to wait for the fonts
to load and ensure the window is always in the same place before
taking a screenshot.

We already merged this in #68, however reverted it in #69 because we
thought it hadn't fixed the problems.

Some more research showed that the CSS problems with text overflow
on mobile are actually a separate bug with the GOV.UK Design System[3].

Therefore we can merge this again and it'll hopefully fix the problems
we can address and allow us to stay up to date with backstop.