Upgrade to BackstopJS 5.3.5
We've had great trouble with the newer versions of Backstop: after upgrading, we see a lot of flakiness due to fonts not rendering.

From my testing, it seems as if the flakiness disappears when I limit BackStopJS to capturing only one thing at a time. This seems to fix the issue without a major impact on run time. Let's see what impact this has when run on Jenkins.

I don't know what the underlying problem is - it could be a local problem, or a problem with the Digital Marketplace itself.